+++ src/app/events/GuildMemberUpdate.ts
@@ -5,6 +5,7 @@ import RemoveGuestPass from '../service/guest-pass/RemoveGuestPass';
 import { DiscordEvent } from '../types/discord/DiscordEvent';
 import ServiceUtils from '../utils/ServiceUtils';
 import sendGuildWelcomeMessage from './welcomeMats/GuildMats';
+import Log, { LogUtils } from '../utils/Log';
 
 export default class implements DiscordEvent {
 	name = 'guildMemberUpdate';
@@ -27,20 +28,19 @@ export default class implements DiscordEvent {
 
 				const removedRoles = oldMember.roles.cache.filter(role => !newMember.roles.cache.has(role.id));
 				if (removedRoles.size > 0) {
-					console.debug(`The roles ${removedRoles.map(r => r.name)} were removed from ${oldMember.displayName}.`);
+					Log.debug(`The roles ${removedRoles.map(r => r.name)} were removed from ${oldMember.displayName}.`);
 					this.handleRolesRemoved(newMember as GuildMember, removedRoles);
 					return;
 				}
 
 				const addedRoles = newMember.roles.cache.filter(role => !oldMember.roles.cache.has(role.id));
 				if (addedRoles.size > 0) {
-					console.debug(`The roles ${addedRoles.map(r => r.name)} were added to ${oldMember.displayName}.`);
+					Log.debug(`The roles ${addedRoles.map(r => r.name)} were added to ${oldMember.displayName}.`);
 					this.handleRolesAdded(newMember as GuildMember, addedRoles);
 				}
 			}
 		} catch (e) {
-			console.error(e);
-			console.error('Retrieving member partial failed');
+			LogUtils.logError('Retrieving member partial failed', e);
 			return;
 		}
 	}
@@ -55,10 +55,10 @@ export default class implements DiscordEvent {
 		roles.each(async role => {
 			switch (role.id) {
 			case roleIds.guestPass:
-				await AddGuestPass(guildMember).catch(err => console.error(err));
+				await AddGuestPass(guildMember).catch(err => LogUtils.logError('failed to add guest pass', err));
 				break;
 			case roleIds.developersGuild:
-				sendGuildWelcomeMessage.devGuildMat(guildMember).catch(err => console.error(err));
+				sendGuildWelcomeMessage.devGuildMat(guildMember).catch(err => LogUtils.logError('failed to send dev guild mat', err));
 				break;
 			}
 		});
@@ -74,7 +74,7 @@ export default class implements DiscordEvent {
 		roles.each(role => {
 			switch (role.id) {
 			case roleIds.guestPass:
-				RemoveGuestPass(guildMember).catch(err => console.error(err));
+				RemoveGuestPass(guildMember).catch(err => LogUtils.logError('failed to remove guest pass', err));
 				break;
 			}
 		});