+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## 1.8.0-SNAPSHOT
 
 1. Fix scoap-squad and update start season 2
+2. Integrate with logDNA sdk, turn off native console.log
+   - fix CODE_OWNERS file
+   - fix `/poap schedule` command
+   - misc enhancements
+   - configure poap time to env variable
+   - set min poap participate time to 10 mins
+   - display guild name and event for poap distribution message
+   - send message on instructions for POAP approval
 
 ## 1.7.2-RELEASE (2021-10-12)